chore(rc): centralize rc logic and add debug logging
